hub/hub/requirements.txt
aiomqtt
python-dotenv
bleak